fixed bug for create order. in-active product cannot be order or checkout

+++ Controllers/order-controllers/createOrderFromCart.js
@@ -27,44 +27,51 @@ module.exports.createOrderFromCart = (request, response)=>{
                 else{
                     Product.findById(data.productId)
                     .then( result => {
-                        if (result.stocks >= data.quantity){
+                        if(!result.isActive){
+                            response.send("This product is not yet active or no available stocks right now!")
+                        }
+                        else{
+                            if (result.stocks >= data.quantity){
                             
 
-                            let newOrder = new Order(
-                                {
-                                    userId: data.userId,
-                                    userEmail: data.userEmail,
-                    
-                                    productId: data.productId,
-                                    productName: data.productName,
-                                    productDescription: data.productDescription,
-                                    price: data.price,
-                                    quantity: data.quantity,
-                                    totalAmount:data.totalAmount
+                                let newOrder = new Order(
+                                    {
+                                        userId: data.userId,
+                                        userEmail: data.userEmail,
+                        
+                                        productId: data.productId,
+                                        productName: data.productName,
+                                        productDescription: data.productDescription,
+                                        price: data.price,
+                                        quantity: data.quantity,
+                                        totalAmount:data.totalAmount
+                                    }
+                                );
+    
+                                result.stocks -= data.quantity;
+                                result.save()
+                                .then(
+                                    //This will delete the data from a cart.
+                                    data.delete()).catch(error=>response.send(error))
+                                
+                
+                                .then( ()=> {
+                
+                                    // This will save new Order to the orders collection
+                                    newOrder.save()
+                                    .then(saveData => response.send(`${saveData.productName} is now on going process. Thank you for your order!`))
+                                    .catch(error => response.send(error))
                                 }
-                            );
-
-                            result.stocks -= data.quantity;
-                            result.save()
-                            .then(
-                                //This will delete the data from a cart.
-                                data.delete()).catch(error=>response.send(error))
-                            
-            
-                            .then( ()=> {
-            
-                                // This will save new Order to the orders collection
-                                newOrder.save()
-                                .then(saveData => response.send(`${saveData.productName} is now on going process. Thank you for your order!`))
+                                )
                                 .catch(error => response.send(error))
+    
                             }
-                            )
-                            .catch(error => response.send(error))
+                            else {
+                                response.send(`We don't have enough stock for your order. Our stocks is:${result.stocks} and your order is:${data.quantity}`)
+                            }
+                        }
 
-                        }
-                        else {
-                            response.send(`We don't have enough stock for your order. Our stocks is:${result.stocks} and your order is:${data.quantity}`)
-                        }
+                       
                         
             
                     })
